Move transit, GBFS and isochrone routes off the /proxy prefix
/proxy is for tile requests MapLibre fetches directly, where the server
strips CORS and keeps provider keys back. These are ordinary authenticated
data endpoints and now live in their own modules, sharing a Barrelman client.
