+++ server/src/controllers/index.ts
@@ -6,6 +6,9 @@ export { default as user } from './user.controller'
 export { default as directions } from './directions.controller'
 export { default as place } from './place.controller'
 export { default as proxy } from './proxy.controller'
+export { default as transit } from './transit.controller'
+export { default as gbfs } from './gbfs.controller'
+export { default as isochrone } from './isochrone.controller'
 export { default as library } from './library'
 export { default as integrations } from './integration.controller'
 export { default as search } from './search.controller'
