+++ server/src/controllers/proxy.controller.ts
@@ -1,110 +1,23 @@
+/**
+ * Tile proxy.
+ *
+ * Map tile sources are fetched by MapLibre itself, which means they can't
+ * attach our auth header and are subject to the tile server's CORS policy.
+ * These routes exist so those requests can go out from the server instead:
+ * they strip CORS out of the equation and keep provider keys server-side.
+ *
+ * That is the *only* thing that belongs here. Ordinary data endpoints that
+ * happen to be served by another service are not proxies — they live in their
+ * own module (see transit, gbfs and isochrone controllers, which all call
+ * Barrelman through `services/barrelman.service.ts`).
+ */
+
 import { Elysia } from 'elysia'
-import { requireAuth } from '../middleware/auth.middleware'
 import { integrationManager } from '../services/integrations'
-import {
-  IntegrationCapabilityId,
-  IntegrationId,
-} from '../types/integration.types'
+import { IntegrationId } from '../types/integration.types'
 import { logError } from '../lib/logger'
-import {
-  ISOCHRONE_MODES,
-  MAX_CONTOURS,
-  maxDurationForMode,
-  type IsochroneMode,
-} from '../types/isochrone.types'
 
 const app = new Elysia({ prefix: '/proxy' })
-
-/** Upstream timeout for everything but isochrones. */
-const DEFAULT_BARRELMAN_TIMEOUT_MS = 10_000
-
-interface BarrelmanProxyOptions {
-  /** Cache-Control on the response we return. Default 'no-cache'. */
-  cacheControl?: string
-  /** Upstream timeout. Default 10s. */
-  timeoutMs?: number
-  /**
-   * Forward the upstream JSON error body instead of replacing it with a
-   * generic message. Only for endpoints whose 4xx bodies say something the
-   * client can act on — isochrone's validation messages name the offending
-   * parameter and its limit, which is worth surfacing.
-   */
-  forwardErrorBody?: boolean
-}
-
-/**
- * Proxy a request to Barrelman. Handles integration config lookup, auth
- * header, error response wrapping, and timeout.
- */
-async function proxyBarrelman(
-  path: string,
-  query: Record<string, any>,
-  options: BarrelmanProxyOptions = {},
-): Promise<Response> {
-  const {
-    cacheControl = 'no-cache',
-    timeoutMs = DEFAULT_BARRELMAN_TIMEOUT_MS,
-    forwardErrorBody = false,
-  } = options
-
-  const systemIntegration = integrationManager
-    .getConfiguredIntegrations()
-    .find((i) => i.integrationId === IntegrationId.BARRELMAN)
-
-  const config = systemIntegration?.config as
-    | { host?: string; apiKey?: string }
-    | undefined
-  if (!config?.host) {
-    return new Response(JSON.stringify({ error: 'Barrelman not configured' }), {
-      status: 501,
-      headers: { 'Content-Type': 'application/json' },
-    })
-  }
-
-  const params = new URLSearchParams()
-  for (const [k, v] of Object.entries(query)) {
-    if (v != null && v !== '') params.set(k, String(v))
-  }
-
-  const headers: Record<string, string> = {}
-  if (config.apiKey) {
-    headers['Authorization'] = `Bearer ${config.apiKey}`
-  }
-
-  const response = await fetch(
-    `${config.host}${path}?${params}`,
-    { headers, signal: AbortSignal.timeout(timeoutMs) },
-  )
-
-  if (!response.ok) {
-    // Return a clean JSON error instead of forwarding upstream HTML. Opting
-    // in to forwardErrorBody passes a JSON body straight through; anything
-    // else upstream sends (an HTML error page, say) still gets replaced.
-    const body = forwardErrorBody ? await safeJson(response) : null
-    return new Response(
-      JSON.stringify(body ?? { error: `Upstream error: ${response.status}` }),
-      { status: response.status, headers: { 'Content-Type': 'application/json' } },
-    )
-  }
-
-  return new Response(await response.arrayBuffer(), {
-    status: 200,
-    headers: {
-      'Content-Type': 'application/json',
-      'Cache-Control': cacheControl,
-    },
-  })
-}
-
-/** Parse a response body as JSON, or null when it isn't JSON at all. */
-async function safeJson(response: Response): Promise<unknown | null> {
-  try {
-    const body = await response.json()
-    return body && typeof body === 'object' ? body : null
-  } catch {
-    return null
-  }
-}
 
 // Helper function to proxy tile requests with integration API key
 async function proxyTileRequest(
@@ -305,174 +218,5 @@ app.get(
     },
   },
 )
-
-// ── Transit proxy endpoints (authenticated) ─────────────────────
-// All transit endpoints require auth to prevent anonymous enumeration
-// of live vehicle positions and route topology.
-const transitProxy = new Elysia({ prefix: '/transit' }).use(requireAuth)
-
-transitProxy.get('/vehicles', ({ query }) =>
-  proxyBarrelman('/transit/vehicles', query, { cacheControl: 'no-cache' }),
-  { detail: { tags: ['Proxy'], summary: 'Proxy GTFS-RT vehicle positions' } },
-)
-
-transitProxy.get('/shapes', ({ query }) =>
-  proxyBarrelman('/transit/shapes', query, { cacheControl: 'public, max-age=86400' }),
-  { detail: { tags: ['Proxy'], summary: 'Proxy route shape geometry' } },
-)
-
-transitProxy.get('/route-vehicles', ({ query }) =>
-  proxyBarrelman('/transit/route-vehicles', query, { cacheControl: 'no-cache' }),
-  { detail: { tags: ['Proxy'], summary: 'Proxy route-specific vehicle positions' } },
-)
-
-transitProxy.get('/trip-stops', ({ query }) =>
-  proxyBarrelman('/transit/trip-stops', query, { cacheControl: 'no-cache' }),
-  { detail: { tags: ['Proxy'], summary: 'Proxy trip stop times' } },
-)
-
-transitProxy.get('/route-detail', ({ query }) =>
-  proxyBarrelman('/transit/route-detail', query, { cacheControl: 'public, max-age=3600' }),
-  { detail: { tags: ['Proxy'], summary: 'Proxy route detail with stops and shape' } },
-)
-
-transitProxy.get('/departures', ({ query }) =>
-  proxyBarrelman('/transit/departures', query, { cacheControl: 'public, max-age=30' }),
-  { detail: { tags: ['Proxy'], summary: 'Proxy upcoming departures at a stop' } },
-)
-
-transitProxy.get('/bikes-allowed', ({ query }) =>
-  proxyBarrelman('/transit/bikes-allowed', query, { cacheControl: 'public, max-age=3600' }),
-  { detail: { tags: ['Proxy'], summary: 'Batch check bikes_allowed for routes' } },
-)
-
-transitProxy.get('/station/:feedId/:stopId', ({ params }) =>
-  proxyBarrelman(
-    `/transit/station/${encodeURIComponent(params.feedId)}/${encodeURIComponent(params.stopId)}`,
-    {},
-    { cacheControl: 'public, max-age=3600' },
-  ),
-  { detail: { tags: ['Proxy'], summary: 'Proxy station detail with entrances and buildings' } },
-)
-
-transitProxy.get('/nearest-entrance', ({ query }) =>
-  proxyBarrelman('/transit/nearest-entrance', query, { cacheControl: 'public, max-age=3600' }),
-  { detail: { tags: ['Proxy'], summary: 'Proxy nearest station entrance lookup' } },
-)
-
-app.use(transitProxy)
-
-// ── GBFS shared mobility proxy ──────────────────────────────────────
-
-// Prefix is relative to `app`, which already contributes `/proxy` — spelling
-// it out here again would mount these at /proxy/proxy/gbfs.
-const gbfsProxy = new Elysia({ prefix: '/gbfs' }).use(requireAuth)
-
-gbfsProxy.get('/nearby-stations', ({ query }) =>
-  proxyBarrelman('/gbfs/nearby-stations', query, { cacheControl: 'no-cache' }),
-  { detail: { tags: ['Proxy'], summary: 'Proxy GBFS nearby stations with availability' } },
-)
-
-gbfsProxy.get('/systems', ({ query }) =>
-  proxyBarrelman('/gbfs/systems', query, { cacheControl: 'public, max-age=3600' }),
-  { detail: { tags: ['Proxy'], summary: 'Proxy GBFS system catalog' } },
-)
-
-app.use(gbfsProxy)
-
-// ── Isochrone proxy (authenticated) ─────────────────────────────────
-
-/**
- * Isochrones are far slower than the rest of Barrelman: a transit contour
- * fans out to hundreds of per-stop GraphHopper searches, and GraphHopper's
- * own per-search timeout is 60s. The shared 10s budget would abort a
- * perfectly healthy request, so this endpoint gets its own.
- */
-const ISOCHRONE_TIMEOUT_MS = 60_000
-
-/**
- * Reject isochrone parameters we can rule out without an upstream round
- * trip. Barrelman validates too, and `forwardErrorBody` surfaces its
- * messages verbatim — this exists to stop the requests that are expensive
- * to discover upstream (a 40-contour transit fan-out) or plainly malformed.
- *
- * Deliberately stricter than Barrelman on one point: only the canonical mode
- * names are accepted, not its aliases ('foot', 'cycling', …). Returns an
- * error message, or null when the query is acceptable.
- */
-function validateIsochroneQuery(query: Record<string, unknown>): string | null {
-  const lat = Number(query.lat)
-  const lng = Number(query.lng)
-  if (!Number.isFinite(lat) || !Number.isFinite(lng)) {
-    return 'lat and lng are required and must be numbers'
-  }
-  if (lat < -90 || lat > 90 || lng < -180 || lng > 180) {
-    return 'lat must be [-90,90], lng must be [-180,180]'
-  }
-
-  const mode = query.mode ? String(query.mode) : 'walk'
-  if (!(ISOCHRONE_MODES as readonly string[]).includes(mode)) {
-    return `Unsupported mode "${mode}". Supported: ${ISOCHRONE_MODES.join(', ')}`
-  }
-
-  if (query.durations) {
-    const durations = String(query.durations).split(',').filter(Boolean)
-    if (durations.length > MAX_CONTOURS) {
-      return `At most ${MAX_CONTOURS} durations per request`
-    }
-    const limit = maxDurationForMode(mode as IsochroneMode)
-    for (const raw of durations) {
-      const seconds = Number(raw)
-      if (!Number.isFinite(seconds) || seconds <= 0) {
-        return `Invalid duration "${raw}" — expected a positive number of seconds`
-      }
-      if (seconds > limit) {
-        return `Duration ${seconds}s exceeds the ${limit}s limit for mode "${mode}"`
-      }
-    }
-  }
-
-  return null
-}
-
-function badRequest(message: string): Response {
-  return new Response(JSON.stringify({ error: message }), {
-    status: 400,
-    headers: { 'Content-Type': 'application/json' },
-  })
-}
-
-// No prefix: `app` already contributes `/proxy`, and mounting the collection
-// root of a prefixed sub-app would hinge on trailing-slash handling.
-const isochroneProxy = new Elysia().use(requireAuth)
-
-isochroneProxy.get('/isochrone', ({ query }) => {
-  const invalid = validateIsochroneQuery(query)
-  if (invalid) return badRequest(invalid)
-
-  return proxyBarrelman('/isochrone', query, {
-    // Transit contours are a function of the departure time, so a cached
-    // answer goes stale in a way the client has no way to notice.
-    cacheControl: 'no-cache',
-    timeoutMs: ISOCHRONE_TIMEOUT_MS,
-    forwardErrorBody: true,
-  })
-}, {
-  detail: {
-    tags: ['Proxy'],
-    summary: 'Proxy reachability polygons for a point',
-    description:
-      'Returns a GeoJSON FeatureCollection with one polygon per requested ' +
-      'contour, ordered smallest first. `durations` is a comma-separated ' +
-      'list of budgets in seconds; `mode` is one of walk, bike, car, transit.',
-  },
-})
-
-isochroneProxy.get('/isochrone/modes', () =>
-  proxyBarrelman('/isochrone/modes', {}, { cacheControl: 'public, max-age=3600' }),
-  { detail: { tags: ['Proxy'], summary: 'Proxy supported isochrone modes and limits' } },
-)
-
-app.use(isochroneProxy)
 
 export default app
